[nrf fromtree] Bluetooth: Host: Amend recycled() callback documentation
The API documentation for the recycled() callback predates [1],
and still warns users to "treat this callback as an ISR",
although it now runs on the system workqueue thread,
as does disconnected().
"Making Bluetooth API calls" to "re-start connectable advertising or
scanning" should no longer be "strongly discouraged".

On the contrary, we can emphasize that this is the right event
to listen for to initiate operations that will try to re-allocate
a freed connection object.

Mention that BT_MAX_CONN configures the size of the connection pool.

Refs:
- [1] efb5d83: Bluetooth: Host: Defer `conn.recycled()` to the syswq

Signed-off-by: Christophe Dufaza <chris@openmarl.org>
(cherry picked from commit 60797818d57756c74cd6cbd4d8de19f6e6d7034a)
Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>