[nrf fromlist] drivers: spi_nrfx_spim: Fix SPIM_REQUESTS_CLOCK() and calls to it
This is a follow-up to commit a0687091713431eda8b02ec737c1c1ad87bfd7da.

As the SPIM_REQUESTS_CLOCK() macro changed the type of its parameter,
all of the calls to it need to be updated accordingly. Also the call to
DT_NODELABEL() needs to be removed from that macro as DT_CLOCKS_CTLR()
already returns a node identifier.

Upstream PR #: 91062

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>