[nrf fromlist] linker: Use zephyr,code-partition value
Enforce usage of values from devicetree if USE_DT_CODE_PARTITION Kconfig
symbol is enabled.

Currently even if this symbol is enabled, the FLASH_LOAD_OFFSET value
is used in the linker script.
The FLASH_LOAD_OFFSET symbol, even being promptless is such case,
may be changed, leading to a very confusing situation, where devicetree
 values have no effect on the build.

A good example of the build that wants to use exactly the same set of
 configs, but a different link address is the A/B update scheme
(known ad Direct XIP in MCUboot), where the same application is linked
to two different addresses.
If the source of the address remains in the Kconifg, the current approach
forces a usage of two different configurations.

Upstream PR #: 91591

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>