[nrf fromlist] soc: nrf54h: don't boot radio core if VTOR is not programmed
Upstream PR #: 91826

Booting the radio core when it is not programmed will typically
cause a reset loop. This can happen when programming multiple
images to a device, and the app core image is programmed before
the radio core.

With this change we avoid the reset loop in that case.

Signed-off-by: Håkon Amundsen <haakon.amundsen@nordicsemi.no>
(cherry picked from commit 7ceb179961d11518bd083958ca524aab2ff01669)