[nrf fromtree] soc: nrf54h: don't boot radio core if VTOR is not programmed
Booting the radio core when it is not programmed will typically
cause a reset loop. This can happen when programming multiple
images to a device, and the app core image is programmed before
the radio core.

With this change we avoid the reset loop in that case.

Signed-off-by: Håkon Amundsen <haakon.amundsen@nordicsemi.no>
(cherry picked from commit 7697eff4fb9470ba1d06e07e00beb9375825f0c5)