[nrf fromlist] tests: boards: nrf: nrfs: Build main image for app core only
Upstream PR #: 91826

Update this multi-core test to always run the `main` and `remote` images
on cpuapp and cpurad respectively.

This is to prepare the test for running with IronSide SE, in which case
keeping cpurad as the main board target wouldn't make as much sense,
because cpurad would have to be started by cpuapp.

Signed-off-by: Grzegorz Swiderski <grzegorz.swiderski@nordicsemi.no>
(cherry picked from commit 921733127636ea9cdb34cc1d74b557bc45029dbb)