[nrf fromtree] soc nrf54lm20a: Fix entropy source
nordic,entropy-prng does not exist in Zephyr (it is part of NCS)
but we have in both the nordic,nrf-cracen-ctrdrbg which is
an actual source of true entropy and works with this SOC.
Let's use that instead.

Fixes failures to build targeting the nrf54lm20dk any test/sample
which uses the entropy driver.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>

Signed-off-by: Sebastian Głąb <sebastian.glab@nordicsemi.no>
(cherry picked from commit 0903efa88221925ba94d9ca12050ddb3be75ef19)