[nrf fromtree] tests: spi_loopback: skip hold_on_cs test if not supported
This early exit section is only reached if there is an error.
In the case of not supported, the ret variable will be 0 and zassert
won't happen. This is because the test was meant to be skipped, not
passed or failed, in this case. So add the skip call.

Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
(cherry picked from commit 0574ac7e845876c8e9fee0c8f3974d519a7ce086)