[nrf noup] boards: xiao_ble: Add static partition manager configuration
The xiao_ble boards ship with a bootloader requiring an app offset of 0x27000.
The upstream board defines this via DT partitions, which will not be used if
partition manager is enabled.
Add a static partition configuration to allow binaries built for this board
to work out-of-the-box in NCS, and match the behavior with sysbuild disabled.

Signed-off-by: Aleksandar Stanoev <aleksandar.stanoev@nordicsemi.no>
(cherry picked from commit 1aff9e5a7abe47398728b6119d4c0d3f2775eb5e)