[nrf noup] Support intermediary SLM optimized modem UART backend
This noup does the following:
- Adds the  modem_backend_uart_slm modem backend which is optimized
  for UART communication to SLM.
- Extends the common modem_cellular device driver to be able to
  switch to include and use the modem_backend_uart_slm backend by
  setting CONFIG_MODEM_CELLULAR_USE_MODEM_BACKEND_UART_SLM. This
  symbol is enabled by default if MODEM_BACKEND_UART_SLM is
  selected.

The noup is intended to live until upstream zephyr has proper
consistent RTS/CTS flow control support across multiple drivers,
and DTR, RTS and CTS support has been added to the upstream modem
UART backends, and lastly, SLM has been updated to use DTR, so
everything is in sync.

Ask Bjarki Arge Andreasen, Markus Lassila or Seppo Takalo when this
noup is to be altered as part of an upmerge for example.

Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>