[nrf noup] modem: backend: uart: SLM optimized modem UART backend
This noup does the following:
- Adds the  modem_backend_uart_slm modem backend which is optimized
  for UART communication to SLM.

The noup is intended to live until upstream zephyr has proper
consistent RTS/CTS flow control support across multiple drivers,
and DTR, RTS and CTS support has been added to the upstream modem
UART backends, and lastly, SLM has been updated to use DTR, so
everything is in sync.

Ask Bjarki Arge Andreasen, Markus Lassila or Seppo Takalo when this
noup is to be altered as part of an upmerge for example.

Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>
Signed-off-by: Markus Lassila <markus.lassila@nordicsemi.no>