[nrf fromlist] modules: nrf_wifi: Fix bustest QSPI crash
Commit 5e25283821a("drivers: wifi: Create dedicated mem pool for Wi-Fi
driver") introduced OSAL dependecy in the Zephyr QSPI driver for HL
read, but in bustest we don't enable nrf_wifi OS module, so, it crashes
here. And we should not be using OSAL APIs in Zephyr code anyway.

And in this case we don't even need to use the heap, so, move the rx
buffer to stack.

Upstream PR #: 93303

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>