[nrf fromtree] net: lib: tls_credentials: add invalid sec tag value
Add a definition for an invalid secure tag value to be used
as a placeholder. Negative values are reserved for internal use,
but some of them should be considered valid.
To be able to check against a value, we need to define an invalid one.

(cherry picked from commit 59a8eebd18eabdd793ad675b3b0befebbf00dddc)

Signed-off-by: Maximilian Deubel <maximilian.deubel@nordicsemi.no>