diag: Fix exiting diag when transmission is ongoing
Crash was observed if "diag stop" was invoked during the transmission.
The issue was sacused by an update to otPlatRadioSleep that made it
compilant with the documentation and not allowing direct Transmeet to
sleep transition.
The transmission result was forwarded to OpenThread making it crash due
to an invalid state.

Signed-off-by: Marek Porwisz <map5-loc@map5-ub.nordicsemi.no>