[nrf fromlist] tests: drivers: comparator: gpio_loopback: Update SNIPPET_ROOT
Upstream PR #: 94796

When SNIPPET is added to the build, it is added to all images. But not
all images have the test directory added to their SNIPPET_ROOT, so we
also update the SNIPPET_ROOT.

This allows, for instance, the uicr image of nrf54H to find this
snippet.

Ref: NCSDK-NONE

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>
(cherry picked from commit e66cc3407d7ebcd6b57bb493a66294617a557640)