[nrf fromtree] tests: drivers: comparator: gpio_loopback: Update SNIPPET_ROOT
When SNIPPET is added to the build, it is added to all images. But not
all images have the test directory added to their SNIPPET_ROOT, so we
also update the SNIPPET_ROOT.

This allows, for instance, the uicr image of nrf54H to find this
snippet.

Ref: NCSDK-NONE

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>
(cherry picked from commit 5f9d20cf38fd9e7a15641b874781d52fb2757cac)