[nrf fromlist] soc: nordic: ironside: Add bootmode service
Added support for the IronSide bootmode service which allows requesting
a reboot into secondary firmware boot mode. In this mode, the secondary
configuration defined in UICR is applied instead of the primary one.

The service provides the ironside_bootmode_secondary_reboot() function
that can pass message data to the boot report of the CPU booted in the
secondary boot mode.

Upstream PR #: 95451

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>