soc: nordic: nrf54h: generate PERIPHCONF macros based on devicetree
Add a script that is used by the build system to generate macros for
populating the PERIPHCONF (global domain peripheral configuration),
based on nodes and properties found in the devicetree.
The script does the same job as nrf-regtool did when building for
nrf54h20 before the move to using IronSide SE, and should make it
so all samples and tests function correctly without workarounds or
manual steps.

The script will also be used on nrf9280, therefore it is placed in
the common uicr directory.

Signed-off-by: Jonathan Nilsen <jonathan.nilsen@nordicsemi.no>