[nrf fromlist] soc: nordic: uicr: Fix dependency issue
Although not reproducible locally, it has been observed in CI that the
uicr image will not always be the last image to be run.

To ensure it is the last image to be run we have it depend on the
'image' image when defined.

The uicr image is generated based on all other images in the build and
must therefore run last.

Upstream PR #: 95785

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>