[nrf noup] boards: nordic: Skip offsets in merged slot
In the merged slot approach, the memory reservation for the MCUboot
header is done in CMake, as it is not obvious, which image includes the
(merged) MCUboot header.
This change will remove the unnecessary gap between images.

Ref: NCSDK-35612

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>
(cherry picked from commit 5f5c5ab8adf8ab33827b1a8b50be771398757699)