drivers: adc: comp: refactor to support power domains and pin ret
The ADC did not implement a PM device structure, this PM could not
resume its power domains. Additionally, the ADC and COMP on the
nrf54h20 both have all of their analog inputs routed to pads in
different power domains than themselves, this means the device
drivers for both of them need to request the pads to be resumed.

To keep the drivers simple, we can rely on HW to keep the power
domain the COMP and ADC are in powered, which happens automatically
when the ADC or COMP is ENABLED. We can then replace their power
domains in the devicetree with the one all of their their analog
inputs need, so PM ensures the pads are powered and not retained
when the ADC or COMP is resumed.

Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>